[s3]: Non exec-files ability to install
Added the ability to install files that are not executable. Files that start
with '.' will be linked. This is a feature for ui-parts of scripts, that whe
want should have .sh extension for editors to colorice them correctly, but we
dont want then to be executed by chance.

Normally shell will hide these files. Perhaps we want the links hidden also?
This is worth to think trough as for example various README files could be added
this way also (which we in such case do want visable).