Added dialog-based UI to ssh_keypair.sh
Old text-based UI is replaced. Old one still works (and is actually
needed), but renamed. To use the old one: .s3..ssh_keypair.sh

New UI is dialog based and requires dialog/xdialog and expect to be
installed to work. If prerequests are not satisfied, function falls back
to old UI.

Old UI works as a back-end and is remote-controlled by a small embedded
Expect-script.

This is a actually really nice example of mixing various different
scriting languages and. Notice the nifty features in the Expect script as
this becomes self-tuned (i.e. code is not finished until called).

Lessons learned:
No scripts in s3/ can depend on s3 to heavily. Writing scripts without the
support of s3-structure is notoriously hard and error-prone. Please avoid!

Category: development