s3.ssh_keypair.sh: Remote port handling added
Added a 3:rd argument to handle the ability to use script with ssh-server
responding on a different port than the default and without the need to
previously manipulate ~/.ssh/config

This is particularly useful for ssh back-doors where host-name to use is
always localhost and the port-number is the identifier. I.e. a situation
where the host-name, even is re-mapped in ~/.ssh/config can't be used as
identifier for ssh_keypair

Change-Id: I06b0915901dd7559b15678548636901f31f80729